requirements.txt
 fastapi==0.115.11
 requests==2.32.3
 uvicorn==0.34.0
 python_dotenv==1.0.1
